Fix: resolve eslint lint errors
