+++ web/src/pages/admin/components/enterprise-feature.tsx
@@ -3,7 +3,7 @@ import { IS_ENTERPRISE } from '../utils';
 export default function EnterpriseFeature({
   children,
 }: {
-  children: () => React.ReactNode;
+  children: React.ReactNode | (() => React.ReactNode);
 }) {
   return IS_ENTERPRISE
     ? typeof children === 'function'
