+++ web/src/pages/admin/forms/import-excel-form.tsx
@@ -42,6 +42,7 @@ export const ImportExcelForm = ({
         <FormField
           control={form.control}
           name="file"
+          // eslint-disable-next-line @typescript-eslint/no-unused-vars
           render={({ field: { onChange, value, ...field } }) => (
             <FormItem>
               <FormLabel className="text-sm font-medium">
